ft_transcendence-main/src/user-web/requirements.txt
Django>=5.0,<6.0
psycopg2>=2.9,<3.0
pillow>=10.2.0
requests>=2.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.5
gunicorn>=21.0.0,<21.1.0
django-otp>=1.3.0,<1.4.0
